Use golang's ioutil.TempFile instead of CUPS' equivalend.
Removes wonky cross-module dependencies.
